added functions for each extract file so they can be called externally

+++ extract/company_extract.py
@@ -1,5 +1,6 @@
 import csv
 import yfinance as yf
+from os.path import exists
 
 
 # Makes use of yfinance and obtains the company country
@@ -10,18 +11,22 @@ def get_data(symbol, name, sector):
     return data
 
 
-print("Pulling data from Yahoo! Finance...")
+def extract_companies():
+    if not exists('../data/company_data.csv'):
+        print("Pulling data from Yahoo! Finance...")
 
-# Writes the obtained data from the source file and yfinance to a new company_data.csv file
-with open('../data/company_data.csv', mode='w', newline='', encoding='utf-8') as file:
-    writer = csv.DictWriter(file, fieldnames=['ticker', 'name', 'sector', 'country'])
-    writer.writeheader()
+        # Writes the obtained data from the source file and yfinance to a new company_data.csv file
+        with open('../data/company_data.csv', mode='w', newline='', encoding='utf-8') as file:
+            writer = csv.DictWriter(file, fieldnames=['ticker', 'name', 'sector', 'country'])
+            writer.writeheader()
 
-    # Reads the original file and obtain the symbol and sector data
-    with open('../data/src/SP500.csv', mode='r', encoding='utf-8') as sp500:
-        reader = csv.DictReader(sp500)
-        for row in reader:
-            row_data = get_data(row['Symbol'], row['Name'], row['Sector'])
-            writer.writerow(row_data)
+            # Reads the original file and obtain the symbol and sector data
+            with open('../data/src/SP500.csv', mode='r', encoding='utf-8') as sp500:
+                reader = csv.DictReader(sp500)
+                for row in reader:
+                    row_data = get_data(row['Symbol'], row['Name'], row['Sector'])
+                    writer.writerow(row_data)
 
-print("File 'company_data.csv' successfully created!")
+        print("File 'company_data.csv' successfully created!")
+    else:
+        print('company_data.csv already exists! Skipping...')
