+++ extract/financial_data_extract.py
@@ -1,5 +1,6 @@
 import csv
 import yfinance as yf
+from os.path import exists
 
 
 # Makes use of yfinance and obtains the company country
@@ -22,20 +23,24 @@ def get_data(symbol):
     return data
 
 
-print("Pulling data from Yahoo! Finance...")
+def extract_financial_data():
+    if not exists('../data/financial_data.csv'):
+        print("Pulling data from Yahoo! Finance...")
 
-# Writes the obtained data from the source file and yfinance to a new company_data.csv file
-with open('../data/financial_data.csv', mode='w', newline='', encoding='utf-8') as file:
-    fieldnames = ['Ticker', 'Date', 'Open', 'Close', 'High', 'Low', 'Volume', 'Dividends', 'Stock Splits']
-    writer = csv.DictWriter(file, fieldnames=fieldnames)
-    writer.writeheader()
+        # Writes the obtained data from the source file and yfinance to a new company_data.csv file
+        with open('../data/financial_data.csv', mode='w', newline='', encoding='utf-8') as file:
+            fieldnames = ['Ticker', 'Date', 'Open', 'Close', 'High', 'Low', 'Volume', 'Dividends', 'Stock Splits']
+            writer = csv.DictWriter(file, fieldnames=fieldnames)
+            writer.writeheader()
 
-    # Reads the original file and obtain the symbol and sector data
-    with open('../data/src/SP500.csv', mode='r', encoding='utf-8') as sp500:
-        reader = csv.DictReader(sp500)
-        for row in reader:
-            row_data = get_data(row['Symbol'])
-            for data_row in row_data:
-                writer.writerow(data_row)
+            # Reads the original file and obtain the symbol and sector data
+            with open('../data/src/SP500.csv', mode='r', encoding='utf-8') as sp500:
+                reader = csv.DictReader(sp500)
+                for row in reader:
+                    row_data = get_data(row['Symbol'])
+                    for data_row in row_data:
+                        writer.writerow(data_row)
 
-print("File 'financial_data_extract.csv' successfully created!")
+        print("File 'financial_data_extract.csv' successfully created!")
+    else:
+        print('financial_data.csv already exists! Skipping...')
